+++ default.nix
@@ -1,1 +1,2 @@
-(import ./release.nix {}).flay
+{ ghc-ver ? "821" }:
+(import ./release.nix {}).${"flay-ghc" + ghc-ver}