+++ shell.nix
@@ -1,1 +1,2 @@
-(import ./default.nix).env
+{ ghc-ver ? "821" }:
+(import ./default.nix { inherit ghc-ver; }).env